requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
feedgen==0.9.0
python-dateutil>=2.8.0
